requirements.txt
fastapi
uvicorn
jinja2
pypdf
python-docx
openai
supabase
pydantic
python-dotenv
httpx
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
